Move static functions out from BaseAlgorithm
